DataBaseHelper: store existing version, set version to current version
Remove some code that was being used to debug. Store currentversion so that the table helpers can do the functions that they will need, but also set the database version to the current version to ensure that there are no conflicts by trying to do this from one of the other helpers by calling incrementDatabaseVersion. For example, if the version has increased more than one number this function would only increase it once causing the other helpers to try and run any code they would need for the various versions multiple times.