StatusBoardFrame: Keylistener time threshold
Implement a configurable time threshold that all 18 characters have to occur within in order to be considered "valid input" from the barcode scanner. This should help prevent problems caused by 'phantom keystrokes'. If a valid character is detected, but the time since the initial key in the sequence was detected is greater than the threshold it will discard the old input and start from scratch with the new character.